# backend-go/internal/api/openapi.v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "翻译助手 langcross OpenAPI",
    "version": "1.0.0",
    "description": "翻译助手（langcross）开放 API v1 规范（与 Python SDK 契约一致）。鉴权：X-API-Key 头（开放 API 专用 Key）。所有计量按积分计费，受租户余额与每日配额约束。"
  },
  "servers": [
    {
      "url": "https://{host}/openapi/v1",
      "variables": {
        "host": {
          "default": "translator.example.com",
          "description": "部署域名（Caddy 已配置 TLS）"
        }
      }
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/balance": {
      "get": {
        "summary": "租户积分余额",
        "operationId": "getBalance",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "balance_points": {
                      "type": "integer"
                    },
                    "balance_sentences_approx": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺失（error_code: invalid_api_key）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "该 API Key 今日调用次数已达上限（error_code: key_quota_exceeded）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "description": "统一错误响应体：success 恒为 false，code 为稳定错误码（前端 i18n 与分支依据），message 为用户可读提示。★ F-64①（批 I-7 2026-09-26）：状态码与 code 同源（internal/errors 的 openAPIStatusByCode 单点决定），不存在「200 + success:false」这一形态",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "code": {
            "type": "string",
            "enum": [
              "invalid_api_key",
              "key_quota_exceeded",
              "forbidden",
              "not_found",
              "bad_request",
              "text_too_long",
              "internal",
              "no_result",
              "task_failed",
              "insufficient_balance",
              "rate_limited",
              "daily_quota_exceeded",
              "rejected",
              "not_ready"
            ]
          },
          "error_code": {
            "type": "string",
            "description": "兼容 < 1.0.4 SDK 的别名字段（与 code 同值）；★ F-64①（批 I-7）明确：code 为正主、error_code 为历史别名，两者由同一处下发，删任一个都会打断在生产的接入方"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object"
          },
          "trace_id": {
            "type": "string",
            "description": "链路追踪 ID，工单排查用"
          }
        }
      }
    }
  }
}